pyproject.toml
[project]
name = "move-it"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-storage>=3.3.1",
    "paramiko>=4.0.0",
    "python-dotenv>=1.1.1",
]
